refactor(session-guide-configurator): trim comments, dedupe copy-link logic
Comment cleanup pass across the app and this PR's changes to shared
files: trimmed narrative comments (decision history, ticket/PLAN.md
references, cross-tool attribution) down to the actual invariant, or
removed them where the code already reads clearly. No behavior changes.

Also collapsed ConfigEditor.js's and Library.js's near-identical
handleCopyLink implementations into one shared copyRowLinkWithToast()
in utils.js.
