+++ event-libs/v1/blocks/sessions-guide/utils/parse-config.js
@@ -1,25 +1,16 @@
 import { detectUserTimezone } from './time.js';
 
-// Legacy fallback for FilterPanel.js, which still filters directly against session
-// fields by id (s[id]) — untouched until it's rewired to consume the ESP-derived
-// authoredFilterCategories shape below (deferred, separate work).
+// Legacy fallback for FilterPanel.js, which indexes sessions by category id, not by
+// ESP attributeId; the authored categories are exposed separately as authoredFilterCategories.
 const DEFAULT_FILTER_CATEGORIES = [
   { id: 'track', label: 'Channel' },
   { id: 'type', label: 'Session Type' },
 ];
 
-// Reads the block's config from the data-session-guide-config attribute that
-// decorate.js's prebuildAutoBlock sets at decoration time (Session Guide
-// Configurator's Copy Link output, decoded there via parseEncodedConfig). There's no
-// authoring-table path anymore — the URL-encoded config is the only source of truth.
-//
-// Only surface/theme/userTz/registerUrl are consumed by the component tree today.
-// headings/behaviorFlags/swimlaneOrder/authoredFilterCategories are carried through
-// as new fields so they're available once DrawerHeader.js/FilterPanel.js/
-// OnDemandView.js are rewired to read them — that rewiring is separate work.
-// filterCategories itself is deliberately left as the legacy track/type default
-// above (not the new authored shape) to avoid silently breaking FilterPanel.js,
-// which indexes sessions by category id, not by ESP attributeId.
+// Config comes solely from the data-session-guide-config attribute decorate.js sets
+// (decoded via parseEncodedConfig); there is no authoring-table path.
+// headings/behaviorFlags/swimlaneOrder/authoredFilterCategories aren't consumed by the
+// component tree yet — keep returning them for callers not yet wired up to read them.
 export function parseSessionsGuideConfig(el, { logPrefix, forcedSurface } = {}) {
   let authored = {};
   try {
