+++ event-libs/v1/services/sessions/sessions-api.js
@@ -366,9 +366,9 @@ export function normalizeSessions(rawSessions) {
 // attribute for its own track editor) doesn't carry a second, independently-drifting copy.
 export const TRACK_ATTRIBUTE_NAME = 'Primary Track for Agenda (Digital Agenda)';
 
-// Generic session/track helpers, not Tier-1-specific — moved here (2026-07-30) so both
-// tier-1-event-configurator and session-guide-configurator import the same
-// implementation instead of one DA app importing from the other's UI code.
+// Generic session/track helpers, not Tier-1-specific — shared here so
+// tier-1-event-configurator and session-guide-configurator use the same implementation
+// instead of one importing from the other's UI code.
 export function getSessionTrack(session) {
   const attr = (session?.customAttributes || []).find((a) => a?.name === TRACK_ATTRIBUTE_NAME);
   return attr?.values?.[0]?.label ?? attr?.values?.[0]?.value ?? null;
@@ -384,13 +384,9 @@ export function extractDistinctTracks(sessions) {
 }
 
 // Derives facetable custom attributes + their distinct values from an already-fetched
-// session catalog. Mirrors the same enabled/inputType/valueId gate ESP's own
-// /session-facets applies server-side (see events-service-platform's
-// resolveCustomAttributes/buildAttributeIndexItems), so results match it field-for-
-// field, without the extra network round-trip — both consumers (session-guide-
-// configurator's own Filters step, and sessions-guide's FilterPanel.js) already have
-// the full session catalog in memory for other reasons. See
-// session-guide-configurator/PLAN.md §7 for the full design writeup.
+// session catalog, mirroring the enabled/inputType/valueId filtering ESP's own
+// /session-facets endpoint applies server-side, so results match it without an extra
+// network round-trip.
 export function deriveFacetableAttributes(sessions) {
   const attributeMap = new Map(); // attributeId -> { attributeId, label, values: Map<valueId, {...}> }
   (sessions || []).forEach((session) => {
